requirements.txt
smtplib
pandas
python-dotenv